corrected module names which are in wrong folders
